Google Analytics
Tracks exercise visits as page views, and button clicks as events.
Will, before release, count the number of attempts a user has made
on a particular exercise.

This should provide very useful insight into the difficulty curve
of the workshop and general usage patterns.

Analytics is enabled by default, but explained alongside the menu
and convient to opt out of.